FLT_K100_TOF: use `csvsimple` for tables
unfortunately, `csvsimple` seems to have some strange issue with letting
multibyte characters (i.e. Chinese) in table headers. possibly because
it is formatted with `\Roman{}`, but I'm too lazy to look into the issue
more. it's easier to just translate the headers than it is to find some
hack to work around it, as it seems embedded within the `csvsimple`
package itself.